requirement.txt
streamlit
googlesearch-python
beautifulsoup4
requests